change quanity and stock level of StockLineItem to double

+++ lib/view/stock/stock.line.item.list.view/stock.line.item.list.view.dart
@@ -47,7 +47,7 @@ class StockLineItemListView extends ConsumerWidget {
                       inputFormatters: <TextInputFormatter>[FilteringTextInputFormatter.digitsOnly],
                       initialValue: e.quantity.toString(),
                       onChanged: (value) {
-                        optionOf(int.tryParse(value)).fold(
+                        optionOf(double.tryParse(value)).fold(
                             () => null,
                             (a) => ref
                                 .read(stockLineItemControllerProvider.notifier)
